% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\name{colnames}
\alias{colnames}
\title{Return column names.}
\usage{
colnames(client)
}
\arguments{
\item{client}{Instance of \code{vantage.infrastructure::Client}}

\item{expl_vars}{\code{Character} vector of variables to include}
}
\value{
\code{data.frame} with one row for each site.
}
\description{
Return column names.
}
