% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSums.R
\name{colSums}
\alias{colSums}
\title{Return column sums.}
\usage{
colSums(client, expl_vars)
}
\arguments{
\item{client}{Instance of \code{vtg::Client}}

\item{expl_vars}{\code{Character} vector of variables to include}
}
\value{
\code{data.frame} with one row for each site.
}
\description{
Return column sums.
}
