% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow.R
\name{nrow}
\alias{nrow}
\title{Return number of rows.}
\usage{
nrow(client)
}
\arguments{
\item{client}{Instance of \code{vtg::Client}}
}
\value{
\code{data.frame} with one row for each site.
}
\description{
Return number of rows.
}
