% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPC_colSums.R
\name{RPC_colSums}
\alias{RPC_colSums}
\title{Return column sums.}
\usage{
RPC_colSums(df, expl_vars)
}
\arguments{
\item{expl_vars}{Variables to include.}
}
\value{
column sums
}
\description{
Return column sums.
}
